Remove useless ShoppingCartLog model.
It wasn't a good idea to have a model to keep track of items inserted
in the shopping cart for every user.

The goal was to prevent a user adding an item to his shopping cart if
sum(items sold) + sum(items currently in shopping carts) >= stock.
The problem is that is also asks to have a timeout on a session, and we
would have possibly lose sales as the timeout could have been reached
anytime and the user's shopping cart could have been emptied even if
there were still items in stock.

Now, we consider that an item is addable to the shopping cart only if
the number of successful sales is inferior to the stock.
Of course, it means that if two users add the same item in the shopping
cart and there's only one available, we could run out of stock.
This is not a major issue for two reasons:
- first, we always have enough items in stocks. When we sell 100 items,
  we have 30 more that we don't sell just in case (return, advertising,
  ...) so selling one or two items more than expected is not really a
  problem ;
- second, because this case will probably be very rare, we could also
  refund the second buyer (and if we give him his money back quickly, we
  won't even be charged for a transaction)