Remove null=True from Charfields
It is considered a bad practice to have a Charfield with null=True. It
is preferable to insert an empty string.

The only reason why a Charfield could be NULL is if you want to play
with the unique attribute. If you need to have a Charfield unique but
optional, set unique=True,null=True (then only one field can have the ""
value, but several can be NULL).